+++ cli/src/commands/diff.rs
@@ -14,6 +14,7 @@
 
 use clap_complete::ArgValueCandidates;
 use clap_complete::ArgValueCompleter;
+use indexmap::IndexSet;
 use itertools::Itertools;
 use jj_lib::copies::CopyRecords;
 use jj_lib::repo::Repo;
@@ -21,8 +22,10 @@ use jj_lib::rewrite::merge_commit_trees;
 use tracing::instrument;
 
 use crate::cli_util::print_unmatched_explicit_paths;
+use crate::cli_util::short_commit_hash;
 use crate::cli_util::CommandHelper;
 use crate::cli_util::RevisionArg;
+use crate::command_error::user_error_with_hint;
 use crate::command_error::CommandError;
 use crate::complete;
 use crate::diff_util::get_copy_records;
@@ -44,23 +47,30 @@ use crate::ui::Ui;
 #[command(mut_arg("ignore_all_space", |a| a.short('w')))]
 #[command(mut_arg("ignore_space_change", |a| a.short('b')))]
 pub(crate) struct DiffArgs {
-    /// Show changes in this revision, compared to its parent(s)
+    /// Show changes in these revisions
     ///
-    /// If the revision is a merge commit, this shows changes *from* the
+    /// If there are multiple revisions, then then total diff for all of them
+    /// will be shown. For example, if you have a linear chain of revisions
+    /// A..D, then `jj diff -r B::D` equals `jj diff --from A --to D`. Multiple
+    /// heads and/or roots are supported, but gaps in the revset are not
+    /// supported (e.g. `jj diff -r 'A|C'` in a linear chain A..C).
+    ///
+    /// If a revision is a merge commit, this shows changes *from* the
     /// automatic merge of the contents of all of its parents *to* the contents
     /// of the revision itself.
     #[arg(
         long,
         short,
-        value_name = "REVSET",
+        value_name = "REVSETS",
+        alias = "revision",
         add = ArgValueCandidates::new(complete::all_revisions)
     )]
-    revision: Option<RevisionArg>,
+    revisions: Option<Vec<RevisionArg>>,
     /// Show changes from this revision
     #[arg(
         long,
         short,
-        conflicts_with = "revision",
+        conflicts_with = "revisions",
         value_name = "REVSET",
         add = ArgValueCandidates::new(complete::all_revisions)
     )]
@@ -69,7 +79,7 @@ pub(crate) struct DiffArgs {
     #[arg(
         long,
         short,
-        conflicts_with = "revision",
+        conflicts_with = "revisions",
         value_name = "REVSET",
         add = ArgValueCandidates::new(complete::all_revisions)
     )]
@@ -95,14 +105,14 @@ pub(crate) fn cmd_diff(
     let repo = workspace_command.repo();
     let fileset_expression = workspace_command.parse_file_patterns(ui, &args.paths)?;
     let matcher = fileset_expression.to_matcher();
-    let resolve_revision = |r: &Option<RevisionArg>| {
-        workspace_command.resolve_single_rev(ui, r.as_ref().unwrap_or(&RevisionArg::AT))
-    };
 
     let from_tree;
     let to_tree;
     let mut copy_records = CopyRecords::default();
     if args.from.is_some() || args.to.is_some() {
+        let resolve_revision = |r: &Option<RevisionArg>| {
+            workspace_command.resolve_single_rev(ui, r.as_ref().unwrap_or(&RevisionArg::AT))
+        };
         let from = resolve_revision(&args.from)?;
         let to = resolve_revision(&args.to)?;
         from_tree = from.tree()?;
@@ -111,14 +121,44 @@ pub(crate) fn cmd_diff(
         let records = get_copy_records(repo.store(), from.id(), to.id(), &matcher)?;
         copy_records.add_records(records)?;
     } else {
-        let to = resolve_revision(&args.revision)?;
-        let parents: Vec<_> = to.parents().try_collect()?;
+        let revision_args = args
+            .revisions
+            .as_deref()
+            .unwrap_or(std::slice::from_ref(&RevisionArg::AT));
+        let revisions_evaluator = workspace_command.parse_union_revsets(ui, revision_args)?;
+        let target_expression = revisions_evaluator.expression();
+        let mut gaps_revset = workspace_command
+            .attach_revset_evaluator(target_expression.connected().minus(target_expression))
+            .evaluate_to_commit_ids()?;
+        if let Some(commit_id) = gaps_revset.next() {
+            return Err(user_error_with_hint(
+                "Cannot diff revsets with gaps in.",
+                format!(
+                    "Revision {} would need to be in the set.",
+                    short_commit_hash(&commit_id?)
+                ),
+            ));
+        }
+        let heads: Vec<_> = workspace_command
+            .attach_revset_evaluator(target_expression.heads())
+            .evaluate_to_commits()?
+            .try_collect()?;
+        let roots: Vec<_> = workspace_command
+            .attach_revset_evaluator(target_expression.roots())
+            .evaluate_to_commits()?
+            .try_collect()?;
+
+        // Collect parents outside of revset to preserve parent order
+        let parents: IndexSet<_> = roots.iter().flat_map(|c| c.parents()).try_collect()?;
+        let parents = parents.into_iter().collect_vec();
         from_tree = merge_commit_trees(repo.as_ref(), &parents)?;
-        to_tree = to.tree()?;
+        to_tree = merge_commit_trees(repo.as_ref(), &heads)?;
 
         for p in &parents {
-            let records = get_copy_records(repo.store(), p.id(), to.id(), &matcher)?;
-            copy_records.add_records(records)?;
+            for to in &heads {
+                let records = get_copy_records(repo.store(), p.id(), to.id(), &matcher)?;
+                copy_records.add_records(records)?;
+            }
         }
     }